split building-hours-helpers apart into individual files
