+++ source/views/building-hours/detail/schedule-table.android.js
@@ -6,10 +6,10 @@
 
 import React from 'react'
 import {View, StyleSheet} from 'react-native'
-import type {DayOfWeekEnumType} from '../types'
 import {Card} from '../../components/card'
 import moment from 'moment-timezone'
-import {isBuildingOpenAtMoment} from '../building-hours-helpers'
+import type {NamedBuildingScheduleType} from '../types'
+import {isScheduleOpenAtMoment, getDayOfWeek} from '../lib'
 import {ScheduleRow} from './schedule-row'
 
 export class ScheduleTable extends React.PureComponent {
@@ -20,7 +20,7 @@ export class ScheduleTable extends React.PureComponent {
 
   render() {
     const {now, schedules} = this.props
-    const dayOfWeek = ((now.format('dd'): any): DayOfWeekEnumType)
+    const dayOfWeek = getDayOfWeek(now)
 
     return (
       <View>
@@ -39,7 +39,7 @@ export class ScheduleTable extends React.PureComponent {
                 isActive={
                   set.isPhysicallyOpen !== false &&
                   schedule.days.includes(dayOfWeek) &&
-                  isBuildingOpenAtMoment(schedule, this.state.now)
+                  isScheduleOpenAtMoment(schedule, this.state.now)
                 }
               />,
             )}
