+++ source/views/building-hours/detail/schedule-table.ios.js
@@ -7,8 +7,8 @@
 import React from 'react'
 import {TableView, Section, Cell} from 'react-native-tableview-simple'
 import moment from 'moment-timezone'
-import type {NamedBuildingScheduleType, DayOfWeekEnumType} from '../types'
-import {isBuildingOpenAtMoment} from '../building-hours-helpers'
+import type {NamedBuildingScheduleType} from '../types'
+import {isScheduleOpenAtMoment, getDayOfWeek} from '../lib'
 import {ScheduleRow} from './schedule-row'
 
 export class ScheduleTable extends React.PureComponent {
@@ -20,7 +20,7 @@ export class ScheduleTable extends React.PureComponent {
 
   render() {
     const {now, schedules, onProblemReport} = this.props
-    const dayOfWeek = ((now.format('dd'): any): DayOfWeekEnumType)
+    const dayOfWeek = getDayOfWeek(now)
 
     return (
       <TableView>
@@ -46,7 +46,7 @@ export class ScheduleTable extends React.PureComponent {
                 isActive={
                   set.isPhysicallyOpen !== false &&
                   schedule.days.includes(dayOfWeek) &&
-                  isBuildingOpenAtMoment(schedule, now)
+                  isScheduleOpenAtMoment(schedule, now)
                 }
               />,
             )}
