+++ source/views/building-hours/lib/__tests__/get-short-building-status.test.js
@@ -1,8 +1,8 @@
 // @flow
-import {getDetailedBuildingStatus} from '../building-hours-helpers'
+import {getShortBuildingStatus} from '../get-short-status'
 import {dayMoment} from './moment.helper'
 
-xit('returns a list of [isOpen, scheduleName, verboseStatus] tuples', () => {
+it('checks a list of schedules to see if any are open', () => {
   let m = dayMoment('Fri 1:00pm')
   let building = {
     name: 'building',
@@ -20,39 +20,7 @@ xit('returns a list of [isOpen, scheduleName, verboseStatus] tuples', () => {
     ],
   }
 
-  const actual = getDetailedBuildingStatus(building, m)
-
-  expect(actual).toBeInstanceOf(Array)
-  expect(actual.length).toEqual(1)
-  expect(typeof actual[0].isActive).toBe('boolean')
-  expect(typeof actual[0].label).toBe('string')
-  expect(typeof actual[0].status).toBe('string')
-
-  expect(actual).toMatchSnapshot()
-})
-
-xit('checks a list of schedules to see if any are open', () => {
-  let m = dayMoment('Fri 1:00pm')
-  let building = {
-    name: 'building',
-    category: '???',
-    breakSchedule: {},
-    schedule: [
-      {
-        title: 'Hours',
-        hours: [
-          {days: ['Mo', 'Tu', 'We', 'Th'], from: '10:30am', to: '12:00am'},
-          {days: ['Fr', 'Sa'], from: '10:30am', to: '2:00am'},
-          {days: ['Su'], from: '10:30am', to: '12:00am'},
-        ],
-      },
-    ],
-  }
-
-  const actual = getDetailedBuildingStatus(building, m)
-  expect(actual).toMatchSnapshot()
-
-  expect(actual[0].isActive).toBe(true)
+  expect(getShortBuildingStatus(building, m)).toBe('Open')
 })
 
 it('handles multiple internal schedules for the same timeframe', () => {
@@ -72,11 +40,7 @@ it('handles multiple internal schedules for the same timeframe', () => {
     ],
   }
 
-  const actual = getDetailedBuildingStatus(building, m)
-  expect(actual).toMatchSnapshot()
-
-  expect(actual[0].isActive).toBe(false)
-  expect(actual[1].isActive).toBe(true)
+  expect(getShortBuildingStatus(building, m)).toBe('Open')
 })
 
 it('handles multiple named schedules for the same timeframe', () => {
@@ -100,15 +64,10 @@ it('handles multiple named schedules for the same timeframe', () => {
     ],
   }
 
-  const actual = getDetailedBuildingStatus(building, m)
-  expect(actual).toMatchSnapshot()
-
-  expect(actual[0].isActive).toBe(false)
-  expect(actual[1].isActive).toBe(false)
-  expect(actual[2].isActive).toBe(true)
+  expect(getShortBuildingStatus(building, m)).toBe('Open')
 })
 
-xit('returns false if none are available for this day', () => {
+it('returns false if none are available for this day', () => {
   let m = dayMoment('Sun 1:00pm')
   let building = {
     name: 'building',
@@ -125,10 +84,7 @@ xit('returns false if none are available for this day', () => {
     ],
   }
 
-  const actual = getDetailedBuildingStatus(building, m)
-  expect(actual).toMatchSnapshot()
-
-  expect(actual[0].isActive).toBe(false)
+  expect(getShortBuildingStatus(building, m)).toBe('Closed')
 })
 
 it('returns false if none are open', () => {
@@ -148,8 +104,5 @@ it('returns false if none are open', () => {
     ],
   }
 
-  const actual = getDetailedBuildingStatus(building, m)
-  expect(actual).toMatchSnapshot()
-
-  expect(actual[0].isActive).toBe(false)
+  expect(getShortBuildingStatus(building, m)).toBe('Closed')
 })
