+++ source/views/building-hours/lib/__tests__/summarize-days.test.js
@@ -1,5 +1,5 @@
 // @flow
-import {summarizeDays} from '../building-hours-helpers'
+import {summarizeDays} from '../summarize-days'
 
 describe('returns a single day if only a single day is given', () => {
   let allDays = [
