+++ source/views/building-hours/lib/constants.js
@@ -1,0 +1,16 @@
+// @flow
+import type {DayOfWeekEnumType} from '../types'
+
+export const CENTRAL_TZ = 'America/Winnipeg'
+export const TIME_FORMAT = 'h:mma'
+export const RESULT_FORMAT = 'LT'
+
+export const daysOfTheWeek: DayOfWeekEnumType[] = [
+  'Mo',
+  'Tu',
+  'We',
+  'Th',
+  'Fr',
+  'Sa',
+  'Su',
+]
