+++ source/views/building-hours/lib/get-day-of-week.js
@@ -1,0 +1,6 @@
+// @flow
+import moment from 'moment-timezone'
+import type {DayOfWeekEnumType} from '../types'
+
+export const getDayOfWeek = (m: moment): DayOfWeekEnumType =>
+  ((m.format('dd'): any): DayOfWeekEnumType)
